feat: confirm/unsubscribe endpoints, weather API integration
- implement GET /api/confirm/{token} endpoint to confirm subscriptions via email link
- implement GET /api/unsubscribe/{token} endpoint for instant unsubscription
- integrate WeatherAPI provider and implement GET /api/weather?city={city}
- refactor handler to support both weather and subscription endpoints
- improve .env and docker-compose for SMTP config
- update main.go to use dependency injection for all services
- ensure email confirmations and unsubscriptions use correct tokens

+++ internal/model/weather.go
@@ -2,6 +2,6 @@ package model
 
 type Weather struct {
 	Temperature float64 `json:"temperature"`
-	Humidity    float64 `json:"humidity"`
+	Humidity    int     `json:"humidity"`
 	Description string  `json:"description"`
 }
